chore(tokens): renames button size tokens
DSY-837
